requirements.txt
aiogram>=3.11,<4.0
aiohttp>=3.9,<3.11
aiohttp-socks>=0.8,<1.0
python-dotenv>=1.0,<2.0
